refactor(lite): clean code fixes for r2.10 codecheck
按 clean code 告警清单修复，三类硬指标（CCN/NLOC/入参）+ lint 抓不到的
语义问题（指针/边界/逻辑/异常/接口契约），覆盖 tools / lite_boost / src /
python / java 五个模块共 35 个文件。机械补全与字面量替换占多数；不可拆
的数据驱动 switch 与会破坏测试可读性的重构留作后续处理，单独记录在
任务目录 CSV 中（不在本 PR 内）。

+++ mindspore-lite/src/extendrt/cxx_api/model_pool/model_pool.cc
@@ -582,7 +582,7 @@ Status ModelPool::CreateSingleWorkerConfig(size_t worker_index, const std::share
     worker_config->config_path = runner_config->GetConfigPath();
   }
   worker_config->context = context;
-  worker_config->worker_id = worker_index;
+  worker_config->worker_id = static_cast<int>(worker_index);
   auto status = SetWorkerModelConfig(&worker_config);
   if (status != kSuccess) {
     return status;
